// src/app/core/api/model/create-milestone-dto.ts
/**
 * Wedding API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreateMilestoneDtoTitle } from './create-milestone-dto-title';


export interface CreateMilestoneDto { 
    title: CreateMilestoneDtoTitle;
    plannedDate: string;
    kind?: CreateMilestoneDto.KindEnum;
    reached?: boolean;
}
export namespace CreateMilestoneDto {
    export const KindEnum = {
        INTERNAL: 'internal',
        GUEST_FACING: 'guest-facing'
    } as const;
    export type KindEnum = typeof KindEnum[keyof typeof KindEnum];
}