// src/app/core/api/model/milestone-list-response-dto.ts
/**
 * Wedding API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MilestoneListResponseDtoItemsInner } from './milestone-list-response-dto-items-inner';


export interface MilestoneListResponseDto { 
    items: Array<MilestoneListResponseDtoItemsInner>;
    count: number;
}